requirements.txt
# 核心框架
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# LLM提供商
openai>=1.0.0
dashscope>=1.17.0  # 通义千问
qianfan>=0.3.0    # 百度文心一言

# Docker支持
docker>=7.0.0

# HTTP请求
httpx>=0.27.0
requests>=2.31.0

# 向量数据库（知识库）
faiss-cpu>=1.7.4
sentence-transformers>=2.3.0

# 工具库
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# 日志和工具
rich>=13.0.0
tqdm>=4.66.0

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0

